Move cache max size config
Moving cache max size config from Alluxio specific configuration
to generic cache configuration.
